+++ modesolver/postprocess/__init__.py
@@ -1,4 +1,6 @@
 from .collocate import collocate
 from .poynting import poynting
+from .group_index import group_index
 from .unfold import unfold
 from .scattering import gsm_step
+from .export import field_to_bitmap, field_to_contours